+++ Code/Common/itkValarrayImageContainer.h
@@ -102,7 +102,7 @@ public:
   
   /** Get the number of elements currently stored in the container. */
   unsigned long Size(void) const
-    { return this->ValarrayType::size(); };
+    { return static_cast<unsigned long>(this->ValarrayType::size()); };
 
   /** Tell the container to allocate enough memory to allow at least
    * as many elements as the size given to be stored.  This is NOT
